requirements.txt
python-dotenv
requests
pandas
plotly
streamlit
nbformat
ipykernel
pytest
pytest-cov
responses
pydeck
geopandas
shapely
moviepy
playwright
imageio
imageio-ffmpeg
numpy
